django_backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
defusedxml==0.7.1
Django==5.1.7
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
packaging==24.2
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.3
tableauserverclient==0.36
typing_extensions==4.12.2
urllib3==2.3.0
whitenoise==6.9.0
